update the bump.main.kts scrip to account for -dev and -alpha versions
